Reorganize registries, purge queues, selmer defaults
This does some big reorganizing of the templates. The registry pages are
in one directory. And the form to requeue a job was put into a central
spot.

This also uses the latest core farmhand code to provide a button for
purging queues.

And I also updated the Selmer configuration so it throws an exception
when tags are missing. This helps to figure out when things are not
being set properly.